requirements.txt
langchain
langchain-openai
langchain-ollama
langchain-mistralai
langgraph
pydantic
python-dotenv
sentence-transformers
fastapi
uvicorn[standard]
